rename src/queue.rs to src/build_queue.rs
